requirements.txt
# ── Nexus AI Backend Dependencies ───────────────────────────────
# Core
Flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
Flask-SQLAlchemy==3.1.1

# Database drivers
pymongo==4.11.3
dnspython==2.7.0
pyodbc==5.2.0
SQLAlchemy==2.0.36

# AI / HTTP
httpx==0.28.1
openai==1.61.1
groq==0.15.0

# Security
bcrypt==4.2.1
google-auth==2.38.0
python-dotenv==1.0.1

# Utilities
Werkzeug==3.1.3
gunicorn==23.0.0

# Digital Twin – Avatar & Voice
Pillow==11.1.0
edge-tts==6.1.9
aiofiles==24.1.0
